+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 * Ignoring texture streaming issues
+* Texture size reporting on Unity 2022.2 and above is now accurate
 
 ## [0.10.0] - 2023-05-03